openmp tutotials

+++ code_samples/lec5/fp_precision.cpp
@@ -7,7 +7,7 @@ const unsigned long PRINT_INTERVAL = 1e+06;   // one million
 
 int main(int argc, char *argv[]) {
 
-  float i = 0; 
+  unsigned long i = 0; 
   int j = 0;
   while (i  <= LOOP_BOUND) {
     // print counter value every millionth iteration
